refactor : 과분리된 객체를 단순화위해 NonNegativeMoves 와 NonNegativeReadyCars 제거
