+++ src/main/java/racinggame/business/CarFactory.java
@@ -4,16 +4,22 @@ import java.util.ArrayList;
 import java.util.List;
 import racinggame.model.Car;
 import racinggame.model.JoinCars;
-import racinggame.model.NonNegativeReadyCars;
 
 public class CarFactory {
     
-    public static JoinCars createCars(NonNegativeReadyCars names) {
+    public static JoinCars createCars(String[] names) {
+        validateCar(names.length);
         List<Car> cars = new ArrayList<>();
-        for (String name : names.names()) {
+        for (String name : names) {
             cars.add(new Car(name));
         }
         return new JoinCars(cars);
+    }
+    
+    private static void validateCar(int cars) {
+        if(cars < 1) {
+            throw new IllegalArgumentException("참가시킬 자동차 수 입력이 생략되거나, 0이하");
+        }
     }
     
 }
