+++ src/main/java/racinggame/business/Process.java
@@ -7,8 +7,8 @@ import racinggame.response.GameResponse;
 public class Process {
     
     public GameResponse run(String[] names, int moves) {
-        JoinCars joinCars = CarFactory.createCars(new NonNegativeReadyCars(names));
-        List<ProgressRecord> progressRecords = new GameMove(new NonNegativeMoves(moves)).moveCar(joinCars);
+        JoinCars joinCars = CarFactory.createCars(names);
+        List<ProgressRecord> progressRecords = new GameMove(moves).moveCar(joinCars);
         WinnerCars gameWinners = progressRecords.getLast().findGameWinners();
         return new GameResponse(new GameResult(progressRecords), gameWinners);
     }
