+++ src/test/java/racinggame/business/GameMoveTest.java
@@ -5,22 +5,32 @@ import static racinggame.business.CarFactory.createCars;
 
 import java.util.List;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.ValueSource;
 import racinggame.model.*;
 
 class GameMoveTest {
-
+    
+    @ParameterizedTest
+    @ValueSource(ints = {0, -1})
+    void moveSetTest(int num) {
+        assertThatThrownBy(() -> {
+            new GameMove(num);
+        }).isInstanceOf(IllegalArgumentException.class)
+            .hasMessage("%s", "이동횟수에 입력이 생략되거나, 0이하");
+    }
 
     @Test
     void ruesSetTest() {
-        assertThat(new GameMove(new NonNegativeMoves(5))).isNotNull();
+        assertThat(new GameMove(5)).isNotNull();
     }
 
     @Test
     void moveCountTest() {
         int moveCount = 5;
         String[] strings = {"pobi", "crong","honux"};
-        GameMove gameMove = new GameMove((new NonNegativeMoves(moveCount)));
-        JoinCars cars = createCars(new NonNegativeReadyCars(strings));
+        GameMove gameMove = new GameMove((moveCount));
+        JoinCars cars = createCars(strings);
         List<ProgressRecord> progressRecords = gameMove.moveCar(cars);
         
         assertThat(progressRecords)
@@ -31,8 +41,8 @@ class GameMoveTest {
     void CarCountTest() {
         int moveCount = 5;
         String[] strings = {"pobi", "crong","honux"};
-        GameMove gameMove = new GameMove((new NonNegativeMoves(moveCount)));
-        JoinCars cars = createCars(new NonNegativeReadyCars(strings));
+        GameMove gameMove = new GameMove((moveCount));
+        JoinCars cars = createCars(strings);
         List<ProgressRecord> progressRecords = gameMove.moveCar(cars);
         
         assertThat(progressRecords)
